Community Matrices

Functions for dealing with community matrices.

Community matrix types

This package currently uses a Matrix{Num} as a community matrix. In the near future this will be changed to use a custom CommunityMatrix type which will include useful metadata besides what a standard matrix does.

EcologicalHypergraphs.community_matrix — Function

community_matrix(hg::EcologicalHypergraph)::Matrix{Num}

Creates a community matrix out of an EcologicalHypergraph. The elements of the matrix are Num allowing symbolic manipulations of the matrix using Symbolics.jl.