Dynamical Hypergraphs

The core of EcologicalHypergraph's dynamical tool set is the @functional_form macro. This macro allows you to assign symbolic functions to each node in a hypergraph along with the values of any parameters. This hypergraph with functional form information can be transformed into a system of ODEs which can be solved using Julia's ecosystem of DE solvers.

EcologicalHypergraphs.@functional_form — Macro

@functional_form node begin

    x -> a*e*x # Forwards function
    x -> a*x   # Backwards function
end a e

The @functional_form macro takes care of creating a symbolic function and adding it to a Node or Vector{Node}. It replaces the placeholder variable x with the variable representing the supplied node and disambiguates the parameters being declared from similarily named parameters on different interactions.

If two functions are supplied as in the above example, the first function is the "forwards function" and represents the amount of biomass addition to the subject of the interaction. the second function is the "backwards function" which represents the amount of biomass being removed from the object of the interaction. The difference between these two functions is the loss from the system.

When this asymmetry is not required, you only need to define one function. In this case both the forwards and backwards function of the node are identical. This is the case for loops, such as growth, where deaths and growth can be collected into the same expression (net growth).

@functional_form node begin

    x -> x*r*(1 - x/k)
end r k

Systems of Differential Equations

These functions are used for converting EcologicalHypergraph objects annotated with functions to systems of differential equations for use in either DifferentialEquations.jl or Symbolics.jl

EcologicalHypergraphs.build_symbolic_system — Function

build_symbolic_system(hg::EcologicalHypergraph)::ODESystem

Takes an EcologicalHypergraph and returns an ODESystem object from Symbolics.jl. The equations in this object all remain in symbolic form.

EcologicalHypergraphs.build_numerical_system — Function

build_symbolic_system(hg::EcologicalHypergraph, tspan::Tuple{Int, Int})::ODESystem

Takes an EcologicalHypergraph and returns an ODEProblem object from DifferentialEquations.jl.