Core Functions

Accessing Species

EcologicalNetworks.species — Function

species(hg::EcologicalHypergraph)::Vector{String}

Returns a vector of all the species represented in an EcologicalHypergraph.

species(hg::EcologicalHypergraph)::Vector{String}

Returns a vector of all the species represented in an edge.

species(n::Node)

Returns the species represented by a node.

Accessing Interactions

EcologicalNetworks.interactions — Function

interactions(hg::EcologicalHypergraph)::Vector{Edge}

Returns a vector of all the edges represented in an EcologicalHypergraph

EcologicalHypergraphs.subject — Function

subject(e::Edge)::Node

Returns the Node playing the role of :subject in an Edge

EcologicalHypergraphs.object — Function

object(e::Edge)::Node

Returns the Node playing the role of :object in an Edge

EcologicalHypergraphs.role — Function

role(n::Node)::Symbol

Returns the role played by a node in it's edge.

Accessing Nodes

EcologicalHypergraphs.nodes — Function

nodes(e::Edge)

Returns a vector of the Nodes in an Edge.

EcologicalHypergraphs.modifiers — Function

modifiers(e::Edge)::Vector{Node}

Returns a vector of all the modifier nodes in an edge.

EcologicalHypergraphs.add_modifier! — Function

add_modifier!(e::Edge, n::Node, role::Symbol = :modifier)::Node

Add a modifier to e representing species spp. The optional role argument defaults to :modifier. Once finished, this function also returns a reference to the node that was added.

Predicates

EcologicalHypergraphs.isloop — Function

isloop(e::Edge)::Bool

Returns true if the edge e is a loop (subject and object species are identical). Returns false otherwise.

EcologicalHypergraphs.contains_species — Function

contains_species(e::Edge, spp::Vector{String}, r::Vector{Symbol})::Bool

Returns true if e contains a species in spp which plays a role in r, otherwise false.

contains_species(e::Edge, sp::String, r::Symbol)::Bool

Returns true if e contains species sp playig role r, otherwise false.

EcologicalHypergraphs.isproducer — Function

isproducer(node::Node, trophic_levels::Dict{String, Float64})::Bool

Returns true if the node node is a producer (trophic level == 1).

EcologicalHypergraphs.isconsumer — Function

isconsumer(node::Node, trophic_levels::Dict{String, Float64})::Bool

Returns true if the node node is a consumer (trophic level > 1.0).

EcologicalHypergraphs.subject_is_consumer — Function

subject_is_consumer(edge::Edge, trophic_levels::Dict{String, Float64})::Bool

Returns true if the subject of the edge edge is a consumer.

EcologicalHypergraphs.subject_is_producer — Function

subject_is_producer(edge::Edge, trophic_levels::Dict{String, Float64})::Bool

Returns true if the subject of the edge edge is a producer.