Types

EcologicalHypergraphs.EcologicalHypergraph — Type

Type represeting an EcologicalHypergraph hypergraph.

Implements an Annotated Hypergraph as described in Chodrow and Mellow (2020)

EcologicalHypergraphs.Node — Type

Represents a node in the annotated hypergraph. Stores the species and the role it plays in it's edge.

EcologicalHypergraphs.Edge — Type

Represents a hyperedge in EcologicalHypergraph.